requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
Babel==2.14.0
billiard==4.2.0
boto3==1.34.117
botocore==1.34.117
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
dj-database-url==2.1.0
Django==5.0.4
django-ace==1.32.4
django-autocomplete-light==3.11.0
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-crispy-forms @ git+https://github.com/tbro2020/django-crispy-forms.git@a314c219044edd0da5f58af2daec3cd36a7043f4
django-currentuser==0.7.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.2
django-json-widget==2.0.1
django-mathfilters==1.0.0
django-money==3.4.1
django-phonenumber-field==7.3.0
django-qr-code==4.0.1
django-slick-reporting==1.3.0
django-storages==1.14.3
django-tinymce==4.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
Faker==24.9.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
jmespath==1.0.1
kombu==5.3.7
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.34
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
py-moneyed==3.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==8.0
redis==5.0.3
s3transfer==0.10.1
segno==1.6.1
sentry-sdk==2.9.0
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.3
